field controller "finished"

+++ src/main/java/io/github/raphael_vi/skillkeeper/repositories/SkillsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SkillsRepository extends JpaRepository<Skill, Integer> {
-    List<Skill> findByFieldId(Long fieldId); // Find all skills by field ID
+    List<Skill> findByFieldId(Integer fieldId); // Find all skills by field ID
     List<Skill> findByFieldName(String fieldName);
 }
